fix(angular): fix imports into nx from ng-add
