+++ packages/linter/package.json
@@ -35,7 +35,6 @@
   "dependencies": {
     "@nrwl/devkit": "file:../devkit",
     "@phenomnomnominal/tsquery": "4.1.1",
-    "nx": "file:../nx",
     "tmp": "~0.2.1",
     "tslib": "^2.3.0"
   },
